Add files via upload
Linked List;
Hard;
***************************
there are two solutions: 
(1) do k-1 operations of merging two lists. I think its the right answer but I don't know why it is TLE,  maybe it's the Python's problem
(2) fetch all the values into a new list, sort it and put it back to one lined-list. I think its time complexity is more the solution 1 in theory. But it is truly passed...
the code here is using solution 2. And for solution 1, just jump to the solution to problem 19.